refactor: reorganize component file structure
