+++ src/pages/home/settings/components/alertButton/AlertButton.tsx
