+++ src/pages/home/settings/components/emptySchoolCard/EmptySchoolCard.tsx
