+++ src/pages/home/settings/components/schoolCard/SchoolCard.css
