+++ src/pages/home/settings/components/schoolCard/SchoolCard.tsx
